hidden_markov_graph.py
import pyro
import pyro.distributions as dist
import torch
from torch.distributions import constraints
from pyro.infer import SVI, Trace_ELBO
from pyro.optim import Adam
from scipy.sparse import coo_matrix

# Constants
N = 160000  # Number of nodes
K = 10      # Number of clusters
d = 5       # Dimensionality of feature space

# Simulated sparse connectivity matrix
row_indices = torch.randint(0, N, (1000000,))
col_indices = torch.randint(0, N, (1000000,))
data = torch.randint(0, 2, (1000000,), dtype=torch.float32)
sparse_matrix = coo_matrix((data, (row_indices, col_indices)), shape=(N, N))

# Sigmoid function
def sigmoid(x):
    return 1 / (1 + torch.exp(-x))

# Model definition
def model(observed_sparse_matrix):
    # Latent cluster assignments
    Z = pyro.sample("Z", dist.Categorical(logits=torch.zeros(N, K)))
    
    # Feature matrix for each cluster
    U = pyro.param("U", torch.randn(K, d), constraint=constraints.real)
    
    # Iterate over the sparse matrix
    for i, j, obs in zip(observed_sparse_matrix.row, observed_sparse_matrix.col, observed_sparse_matrix.data):
        U_i = U[Z[i]]
        U_j = U[Z[j]]
        mu_ij = sigmoid((U_i * U_j).sum())
        pyro.sample(f"W_{i}_{j}", dist.Bernoulli(mu_ij), obs=torch.tensor(obs))

# Guide definition
def guide(observed_sparse_matrix):
    # Learnable parameters for variational distribution over Z
    q_logits = pyro.param("q_logits", torch.randn(N, K), constraint=constraints.simplex)
    pyro.sample("Z", dist.Categorical(logits=q_logits))

# Sparse matrix processing
observed_matrix = coo_matrix(sparse_matrix)

# Set up the optimizer and inference
optimizer = Adam({"lr": 0.01})
svi = SVI(model, guide, optimizer, loss=Trace_ELBO())

# Training loop
num_steps = 1000
for step in range(num_steps):
    loss = svi.step(observed_matrix)
    if step % 100 == 0:
        print(f"Step {step} - Loss: {loss:.4f}")

# Output the learned cluster assignments
q_logits = pyro.param("q_logits")
Z_posterior = torch.argmax(q_logits, dim=1)
print("Inferred cluster assignments:", Z_posterior)