requirements.txt
pyro-ppl
scipy
pandas